E_commerce/E_commerce/requirement.txt
django
psycopg2-binary
djangorestframework
django-jazzmin
python-dotenv
